requirements.txt
streamlit
qdrant-client
sentence-transformers
requests
python-dotenv
openai
numpy
pandas
plotly
httpx
pydantic
langchain
langchain-community
tiktoken
